Revert "cham projector fixes/rewrite (#27111)"
This reverts commit 526005615684b7da1e24ea88b1225c34a3c5996f.
