+++ Content.Shared/Polymorph/Components/ChameleonDisguiseComponent.cs
@@ -1,4 +1,3 @@
-using Content.Shared.Polymorph.Systems;
 using Robust.Shared.GameStates;
 using Robust.Shared.Prototypes;
 
@@ -8,22 +7,9 @@ namespace Content.Shared.Polymorph.Components;
 /// Component added to disguise entities.
 /// Used by client to copy over appearance from the disguise's source entity.
 /// </summary>
-[RegisterComponent, NetworkedComponent, Access(typeof(SharedChameleonProjectorSystem))]
-[AutoGenerateComponentState(true)]
+[RegisterComponent, NetworkedComponent, AutoGenerateComponentState(true)]
 public sealed partial class ChameleonDisguiseComponent : Component
 {
-    /// <summary>
-    /// The user of this disguise.
-    /// </summary>
-    [DataField]
-    public EntityUid User;
-
-    /// <summary>
-    /// The projector that created this disguise.
-    /// </summary>
-    [DataField]
-    public EntityUid Projector;
-
     /// <summary>
     /// The disguise source entity for copying the sprite.
     /// </summary>
