+++ Content.Shared/Polymorph/Components/ChameleonProjectorComponent.cs
@@ -1,3 +1,4 @@
+using Content.Shared.Polymorph;
 using Content.Shared.Polymorph.Systems;
 using Content.Shared.Whitelist;
 using Robust.Shared.Prototypes;
@@ -25,26 +26,22 @@ public sealed partial class ChameleonProjectorComponent : Component
     public EntityWhitelist? Blacklist;
 
     /// <summary>
-    /// Disguise entity to spawn and use.
+    /// Polymorph configuration for the disguise entity.
     /// </summary>
     [DataField(required: true)]
-    public EntProtoId DisguiseProto = string.Empty;
+    public PolymorphConfiguration Polymorph = new();
 
     /// <summary>
     /// Action for disabling your disguise's rotation.
     /// </summary>
     [DataField]
     public EntProtoId NoRotAction = "ActionDisguiseNoRot";
-    [DataField]
-    public EntityUid? NoRotActionEntity;
 
     /// <summary>
     /// Action for anchoring your disguise in place.
     /// </summary>
     [DataField]
     public EntProtoId AnchorAction = "ActionDisguiseAnchor";
-    [DataField]
-    public EntityUid? AnchorActionEntity;
 
     /// <summary>
     /// Minimum health to give the disguise.
@@ -59,10 +56,16 @@ public sealed partial class ChameleonProjectorComponent : Component
     public float MaxHealth = 100f;
 
     /// <summary>
-    /// User currently disguised by this projector, if any
+    /// Popup shown to the user when they try to disguise as an invalid entity.
     /// </summary>
     [DataField]
-    public EntityUid? Disguised;
+    public LocId InvalidPopup = "chameleon-projector-invalid";
+
+    /// <summary>
+    /// Popup shown to the user when they disguise as a valid entity.
+    /// </summary>
+    [DataField]
+    public LocId SuccessPopup = "chameleon-projector-success";
 
     /// <summary>
     /// If this disguises when used as an item or is an action.
@@ -75,15 +78,4 @@ public sealed partial class ChameleonProjectorComponent : Component
 
     [DataField]
     public EntityUid? DisguiseActionEntity;
-
-}
-
-public sealed partial class ChameleonDisguisedEvent : EntityEventArgs
-{
-    public EntityUid Disguise;
-
-    public ChameleonDisguisedEvent(EntityUid disguise)
-    {
-        Disguise = disguise;
-    }
 }
