+++ Resources/Locale/en-US/chameleon-projector/chameleon-projector.ftl
@@ -1,4 +1,2 @@
-chameleon-projector-inside-container = There's no room to scan that!
 chameleon-projector-invalid = You can't disguise as that!
 chameleon-projector-success = Projected new disguise.
-chameleon-projector-set-disguise = Set Disguise
