+++ Resources/Prototypes/Entities/Mobs/Player/morph.yml
@@ -117,11 +117,13 @@
     blacklist:
       components:
       - MorphDisguise
+      - InsideEntityStorage # no clark kent going in phone booth and becoming superman
       - Pda # PDAs currently make you invisible /!\
       tags:
       - Wall
       - Structure
-    disguiseProto: MorphDisguise
+    polymorph:
+      entity: MorphDisguise
   - type: GuideHelp
     guides:
     - MinorAntagonists
@@ -139,10 +141,12 @@
   # so it can take damage
   # projector system sets health to be proportional to mass
   - type: Damageable
-  - type: Transform
-    noRot: true # players rotation and anchor is used instead
-  - type: InteractionOutline
-  - type: Clickable
+  - type: MobState
+  - type: MobThresholds
+    thresholds:
+      0: Alive
+      1: Critical
+      200: Dead
   - type: ChameleonDisguise
   - type: InventorySlots
