+++ Resources/Prototypes/Entities/Objects/Devices/chameleon_projector.yml
@@ -17,12 +17,15 @@
     blacklist:
       components:
       - ChameleonDisguise # no becoming kleiner
+      - InsideEntityStorage # no clark kent going in phone booth and becoming superman
       - MindContainer # no
       - Pda # PDAs currently make you invisible /!\
-    disguiseProto: ChameleonDisguise
+    polymorph:
+      entity: ChameleonDisguise
 
 - type: entity
   categories: [ HideSpawnMenu ]
+  parent: BaseMob
   id: ChameleonDisguise
   name: Urist McKleiner
   components:
@@ -30,11 +33,20 @@
   - type: Sprite
     sprite: /Textures/Mobs/Species/Human/parts.rsi
     state: full
-  - type: Transform
-    noRot: true # players rotation and anchor is used instead
-  - type: InteractionOutline
-  - type: Clickable
+  # so people can attempt to pick it up
+  - type: Item
+  # so it can take damage
+  # projector system sets health to be proportional to mass
   - type: Damageable
+  - type: MobState
+  - type: MobThresholds
+    thresholds:
+      0: Alive
+      1: Critical
+      200: Dead
+  - type: MovementSpeedModifier
+    baseWalkSpeed: 1 # precise movement for the perfect spot
+    baseSprintSpeed: 5 # the jig is up
   - type: ChameleonDisguise
 
 # actions
@@ -45,7 +57,6 @@
   components:
   - type: InstantAction
     icon: Interface/VerbIcons/refresh.svg.192dpi.png
-    itemIconStyle: BigAction
     event: !type:DisguiseToggleNoRotEvent
 
 - type: entity
@@ -57,5 +68,4 @@
     icon:
       sprite: Objects/Tools/wrench.rsi
       state: icon
-    itemIconStyle: BigAction
     event: !type:DisguiseToggleAnchoredEvent
